+++ trunk/comet-ms/CometSearch/ThreadPool.h
@@ -192,7 +192,7 @@ public:
    int NumParamsQueued()
    {
       Threading::LockMutex(_poolAccessMutex);
-      int numParamsQueued = _params.size();
+      int numParamsQueued = (int)_params.size();
       Threading::UnlockMutex(_poolAccessMutex);
       return numParamsQueued;
    }
@@ -200,7 +200,7 @@ public:
    int NumActiveThreads()
    {
       Threading::LockMutex(_poolAccessMutex);
-      int numActiveThreads = _numCurrThreads - _threads.size();
+      int numActiveThreads = _numCurrThreads - (int)_threads.size();
       Threading::UnlockMutex(_poolAccessMutex);
       return numActiveThreads;
    }
